requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
appdirs==1.4.4
    # via black
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via
    #   automat
    #   black
    #   service-identity
    #   twisted
automat==20.2.0
    # via twisted
black==20.8b0
    # via -r requirements.in
boto3==1.18.58
    # via -r requirements.in
botocore==1.21.58
    # via
    #   boto3
    #   s3transfer
certifi==2021.5.30
    # via requests
cffi==1.14.6
    # via cryptography
charset-normalizer==2.0.6
    # via requests
click==7.0
    # via
    #   -r requirements.in
    #   black
    #   shub
constantly==15.1.0
    # via twisted
cryptography==3.4.8
    # via
    #   pyopenssl
    #   scrapy
    #   service-identity
cssselect==1.1.0
    # via
    #   parsel
    #   scrapy
dataclasses==0.6
    # via -r requirements.in
django==3.2.8
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
docker==5.0.2
    # via shub
h2==3.2.0
    # via
    #   scrapy
    #   twisted
hpack==3.0.0
    # via h2
hyperframe==5.2.0
    # via h2
hyperlink==21.0.0
    # via twisted
idna==3.2
    # via
    #   hyperlink
    #   requests
incremental==21.3.0
    # via twisted
itemadapter==0.4.0
    # via
    #   itemloaders
    #   scrapy
itemloaders==1.0.4
    # via scrapy
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
    #   itemloaders
lxml==4.6.3
    # via
    #   parsel
    #   scrapy
mypy-extensions==0.4.3
    # via black
parsel==1.6.0
    # via
    #   itemloaders
    #   scrapy
pathspec==0.9.0
    # via black
priority==1.3.0
    # via twisted
protego==0.1.16
    # via scrapy
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.2.8
    # via service-identity
pycparser==2.20
    # via cffi
pydispatcher==2.0.5
    # via scrapy
pyopenssl==20.0.1
    # via scrapy
python-dateutil==2.8.2
    # via botocore
pytz==2021.3
    # via django
pyyaml==5.1.2
    # via
    #   -r requirements.in
    #   shub
queuelib==1.6.2
    # via scrapy
regex==2021.8.28
    # via black
requests==2.26.0
    # via
    #   docker
    #   scrapinghub
    #   shub
retrying==1.3.3
    # via
    #   scrapinghub
    #   shub
s3transfer==0.5.0
    # via
    #   -r requirements.in
    #   boto3
scrapinghub==2.3.1
    # via shub
scrapy==2.5.0
    # via -r requirements.in
service-identity==21.1.0
    # via scrapy
shub==2.13.0
    # via -r requirements.in
six==1.16.0
    # via
    #   automat
    #   parsel
    #   protego
    #   pyopenssl
    #   python-dateutil
    #   retrying
    #   scrapinghub
    #   service-identity
    #   shub
    #   w3lib
    #   websocket-client
sqlparse==0.4.2
    # via django
toml==0.10.2
    # via
    #   black
    #   shub
tqdm==4.55.1
    # via shub
twisted[http2]==21.7.0
    # via scrapy
typed-ast==1.4.3
    # via black
typing-extensions==3.10.0.2
    # via
    #   black
    #   twisted
urllib3==1.25.11
    # via
    #   -r requirements.in
    #   botocore
    #   requests
w3lib==1.22.0
    # via
    #   itemloaders
    #   parsel
    #   scrapy
websocket-client==0.54.0
    # via
    #   -r requirements.in
    #   docker
zope.interface==5.4.0
    # via
    #   scrapy
    #   twisted

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
